+++ lib/src/dialogs/dialog.dart
@@ -1,12 +1,44 @@
 import 'package:arna/arna.dart';
 
+/// An Arna-styled alert dialog.
+///
+/// An alert dialog informs the user about situations that require
+/// acknowledgement.
+///
+/// Typically passed as the child widget to [showArnaDialog], which displays
+/// the dialog.
+///
+/// {@tool snippet}
+///
+/// This snippet shows a method in a [State] which, when called, displays a
+/// dialog box and returns a [Future] that completes when the dialog is
+/// dismissed.
+///
+/// ```dart
+/// Future<void> _showMyDialog() async {
+///   return showArnaDialog<void>(
+///     context: context,
+///     barrierDismissible: false, // user must tap button!
+///     dialog: ArnaAlertDialog(
+///       title: "Title",
+///       message: "Message",
+///       primary: ArnaTextButton(
+///       label: "OK",
+///       onPressed: Navigator.of(context).pop,
+///     ),
+///   );
+/// }
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [ArnaPopupDialog]
+///  * [showArnaDialog]
 class ArnaAlertDialog extends StatelessWidget {
-  final String? title;
-  final String? message;
-  final Widget primary;
-  final Widget secondary;
-  final Widget tertiary;
-
+  /// Creates an alert dialog.
+  ///
+  /// Typically used in conjunction with [showArnaDialog].
   const ArnaAlertDialog({
     Key? key,
     this.title,
@@ -15,6 +47,26 @@ class ArnaAlertDialog extends StatelessWidget {
     this.secondary = const SizedBox.shrink(),
     this.tertiary = const SizedBox.shrink(),
   }) : super(key: key);
+
+  /// The (optional) title of the dialog is displayed in a large font at the top
+  /// of the dialog.
+  final String? title;
+
+  /// The (optional) content of the dialog is displayed in the center of the
+  /// dialog in a lighter font.
+  final String? message;
+
+  /// The primary action that is displayed at the bottom of the
+  /// dialog.
+  final Widget primary;
+
+  /// The (optional) secondary action that is displayed at the bottom of the
+  /// dialog.
+  final Widget secondary;
+
+  /// The (optional) tertiary action that is displayed at the bottom of the
+  /// dialog.
+  final Widget tertiary;
 
   @override
   Widget build(BuildContext context) {
@@ -77,6 +129,8 @@ class ArnaAlertDialog extends StatelessWidget {
                               child: Text(
                                 message!,
                                 maxLines: 5,
+                                style:
+                                    ArnaTheme.of(context).textTheme.textStyle,
                                 textAlign: TextAlign.left,
                               ),
                             ),
@@ -109,12 +163,47 @@ class ArnaAlertDialog extends StatelessWidget {
   }
 }
 
+/// Displays a dialog above the current contents of the app.
+///
+///
+/// This function takes a [dialog] which is used to build the dialog widget.
+/// Content below the dialog is dimmed with a [ModalBarrier].
+///
+/// The [context] argument is used to look up the [Navigator] for the
+/// dialog. It is only used when the method is called. Its corresponding widget
+/// can be safely removed from the tree before the dialog is closed.
+///
+/// The [useRootNavigator] argument is used to determine whether to push the
+/// dialog to the [Navigator] furthest from or nearest to the given `context`.
+/// By default, [useRootNavigator] is true and the dialog route created by
+/// this method is pushed to the root navigator.
+///
+/// If the application has multiple [Navigator] objects, it may be necessary to
+/// call `Navigator.of(context, rootNavigator: true).pop(result)` to close the
+/// dialog rather than just `Navigator.pop(context, result)`.
+///
+/// The [barrierDismissible] argument is used to determine whether this route
+/// can be dismissed by tapping the modal barrier. This argument defaults
+/// to false. If [barrierDismissible] is true, a non-null [barrierLabel] must be
+/// provided.
+///
+/// The [barrierLabel] argument is the semantic label used for a dismissible
+/// barrier. This argument defaults to `null`.
+///
+/// The [barrierColor] argument is the color used for the modal barrier. This
+/// argument defaults to [ArnaColors.barrierColor].
+///
+/// The [routeSettings] will be used in the construction of the dialog's route.
+/// See [RouteSettings] for more details.
+///
+/// Returns a [Future] that resolves to the value (if any) that was passed to
+/// [Navigator.pop] when the dialog was closed.
 Future<T?> showArnaDialog<T>({
   required BuildContext context,
-  required WidgetBuilder builder,
+  required ArnaAlertDialog dialog,
   bool barrierDismissible = false,
   Color barrierColor = ArnaColors.barrierColor,
-  String? barrierLabel = "label",
+  String? barrierLabel,
   bool useRootNavigator = true,
   RouteSettings? routeSettings,
 }) {
@@ -125,7 +214,7 @@ Future<T?> showArnaDialog<T>({
     barrierDismissible: barrierDismissible,
     transitionDuration: Styles.basicDuration,
     routeSettings: routeSettings,
-    pageBuilder: (context, animation, secondaryAnimation) => builder(context),
+    pageBuilder: (context, animation, secondaryAnimation) => dialog,
     transitionBuilder: (_, anim, __, child) => phone(context)
         ? SlideTransition(
             position: Tween(
@@ -133,23 +222,14 @@ Future<T?> showArnaDialog<T>({
               end: Offset.zero,
             ).animate(anim),
             child: FadeTransition(
-              opacity: CurvedAnimation(
-                parent: anim,
-                curve: Styles.basicCurve,
-              ),
+              opacity: CurvedAnimation(parent: anim, curve: Styles.basicCurve),
               child: child,
             ),
           )
         : ScaleTransition(
-            scale: CurvedAnimation(
-              parent: anim,
-              curve: Styles.basicCurve,
-            ),
+            scale: CurvedAnimation(parent: anim, curve: Styles.basicCurve),
             child: FadeTransition(
-              opacity: CurvedAnimation(
-                parent: anim,
-                curve: Styles.basicCurve,
-              ),
+              opacity: CurvedAnimation(parent: anim, curve: Styles.basicCurve),
               child: child,
             ),
           ),