+++ lib/src/dialogs/popup_dialog.dart
@@ -1,21 +1,66 @@
 import 'package:arna/arna.dart';
 import 'package:flutter/cupertino.dart' show CupertinoPageRoute;
 
-class ArnaPopupDialog extends StatelessWidget {
-  final String? title;
-  final Widget headerBarLeading;
-  final Widget headerBarTrailing;
-  final ArnaSearchField? searchField;
-  final Widget body;
-
-  const ArnaPopupDialog({
+/// An Arna-styled popup dialog.
+///
+/// A popup dialog displays a dialog box in desktop and tablet mode.
+///
+///
+/// Typically passed as the child widget to [showArnaDialog], which displays
+/// the dialog.
+///
+/// {@tool snippet}
+///
+/// This snippet shows an ArnaIconButton which, when pressed, displays a dialog
+/// box in desktop and tablet mode and a new page on phones.
+///
+/// ```dart
+/// ArnaIconButton(
+///     icon: Icons.info_outlined,
+///     onPressed: () => showArnaPopupDialog(
+///         context: context,
+///         title: "Title",
+///         body: Container(),
+///     ),
+/// );
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [ArnaAlertDialog]
+///  * [showArnaPopupDialog]
+class _ArnaPopupDialog extends StatelessWidget {
+  /// Creates a popup dialog.
+  ///
+  /// Typically used in conjunction with [showArnaPopupDialog].
+  const _ArnaPopupDialog({
     Key? key,
+    this.headerBarLeading,
     this.title,
-    this.headerBarLeading = const SizedBox.shrink(),
-    this.headerBarTrailing = const SizedBox.shrink(),
+    this.headerBarTrailing,
     this.searchField,
+    this.banner,
     required this.body,
   }) : super(key: key);
+
+  /// The leading widget laid out within the header bar.
+  final Widget? headerBarLeading;
+
+  /// The title displayed in the header bar.
+  final String? title;
+
+  /// The trailing widget laid out within the header bar.
+  final Widget? headerBarTrailing;
+
+  /// The [ArnaSearchField] of the popup dialog.
+  final ArnaSearchField? searchField;
+
+  /// The [ArnaBanner] of the popup dialog.
+  final ArnaBanner? banner;
+
+  /// The body widget of the popup dialog.
+  final Widget body;
 
   @override
   Widget build(BuildContext context) {
@@ -57,7 +102,7 @@ class ArnaPopupDialog extends StatelessWidget {
                   title: title,
                   headerBarTrailing: Row(
                     children: [
-                      headerBarTrailing,
+                      if (headerBarTrailing != null) headerBarTrailing!,
                       ArnaTextButton(
                         label: "Close",
                         onPressed: Navigator.of(context).pop,
@@ -65,6 +110,7 @@ class ArnaPopupDialog extends StatelessWidget {
                     ],
                   ),
                   searchField: searchField,
+                  banner: banner,
                   body: body,
                 ),
               ),
@@ -76,21 +122,66 @@ class ArnaPopupDialog extends StatelessWidget {
   }
 }
 
-class ArnaPopupPage extends StatelessWidget {
-  final String? title;
-  final Widget headerBarLeading;
-  final Widget headerBarTrailing;
-  final ArnaSearchField? searchField;
-  final Widget body;
-
-  const ArnaPopupPage({
+/// An Arna-styled popup dialog.
+///
+/// A popup dialog displays a page in phone mode.
+///
+///
+/// Typically passed as the child widget to [showArnaDialog], which displays
+/// the dialog.
+///
+/// {@tool snippet}
+///
+/// This snippet shows an ArnaIconButton which, when pressed, displays a dialog
+/// box in desktop and tablet mode and a new page on phones.
+///
+/// ```dart
+/// ArnaIconButton(
+///     icon: Icons.info_outlined,
+///     onPressed: () => showArnaPopupDialog(
+///         context: context,
+///         title: "Title",
+///         body: Container(),
+///     ),
+/// );
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [ArnaAlertDialog]
+///  * [showArnaPopupDialog]
+class _ArnaPopupPage extends StatelessWidget {
+  /// Creates a popup dialog page.
+  ///
+  /// Typically used in conjunction with [showArnaPopupDialog].
+  const _ArnaPopupPage({
     Key? key,
+    this.headerBarLeading,
     this.title,
-    this.headerBarLeading = const SizedBox.shrink(),
-    this.headerBarTrailing = const SizedBox.shrink(),
+    this.headerBarTrailing,
     this.searchField,
+    this.banner,
     required this.body,
   }) : super(key: key);
+
+  /// The leading widget laid out within the header bar.
+  final Widget? headerBarLeading;
+
+  /// The title displayed in the header bar.
+  final String? title;
+
+  /// The trailing widget laid out within the header bar.
+  final Widget? headerBarTrailing;
+
+  /// The [ArnaSearchField] of the popup dialog page.
+  final ArnaSearchField? searchField;
+
+  /// The [ArnaBanner] of the popup dialog page.
+  final ArnaBanner? banner;
+
+  /// The body widget of the popup dialog page.
+  final Widget body;
 
   @override
   Widget build(BuildContext context) {
@@ -101,23 +192,60 @@ class ArnaPopupPage extends StatelessWidget {
             icon: Icons.arrow_back_outlined,
             onPressed: () => Navigator.pop(context),
           ),
-          headerBarLeading,
+          if (headerBarLeading != null) headerBarLeading!,
         ],
       ),
       title: title,
       headerBarTrailing: headerBarTrailing,
       searchField: searchField,
+      banner: banner,
       body: body,
     );
   }
 }
 
+/// Displays a dialog in desktop and tablet mode and a new page on phones.
+///
+///
+/// In desktop and tablet mode content below the dialog is dimmed with a
+/// [ModalBarrier].
+///
+/// The [context] argument is used to look up the [Navigator] for the
+/// dialog. It is only used when the method is called. Its corresponding widget
+/// can be safely removed from the tree before the dialog is closed.
+///
+/// The [useRootNavigator] argument is used to determine whether to push the
+/// dialog to the [Navigator] furthest from or nearest to the given `context`.
+/// By default, [useRootNavigator] is true and the dialog route created by
+/// this method is pushed to the root navigator.
+///
+/// If the application has multiple [Navigator] objects, it may be necessary to
+/// call `Navigator.of(context, rootNavigator: true).pop(result)` to close the
+/// dialog rather than just `Navigator.pop(context, result)`.
+///
+/// The [barrierDismissible] argument is used to determine whether this route
+/// can be dismissed by tapping the modal barrier. This argument defaults
+/// to false. If [barrierDismissible] is true, a non-null [barrierLabel] must be
+/// provided.
+///
+/// The [barrierLabel] argument is the semantic label used for a dismissible
+/// barrier. This argument defaults to `null`.
+///
+/// The [barrierColor] argument is the color used for the modal barrier. This
+/// argument defaults to [ArnaColors.barrierColor].
+///
+/// The [routeSettings] will be used in the construction of the dialog's route.
+/// See [RouteSettings] for more details.
+///
+/// Returns a [Future] that resolves to the value (if any) that was passed to
+/// [Navigator.pop] when the dialog was closed.
 Future<T?> showArnaPopupDialog<T>({
   required BuildContext context,
+  Widget? headerBarLeading,
   String? title,
-  Widget headerBarLeading = const SizedBox.shrink(),
-  Widget headerBarTrailing = const SizedBox.shrink(),
+  Widget? headerBarTrailing,
   ArnaSearchField? searchField,
+  ArnaBanner? banner,
   required Widget body,
   bool barrierDismissible = false,
   Color barrierColor = ArnaColors.barrierColor,
@@ -128,11 +256,12 @@ Future<T?> showArnaPopupDialog<T>({
   return phone(context)
       ? Navigator.of(context).push(
           CupertinoPageRoute(
-            builder: (context) => ArnaPopupPage(
-              title: title,
+            builder: (context) => _ArnaPopupPage(
               headerBarLeading: headerBarLeading,
+              title: title,
               headerBarTrailing: headerBarTrailing,
               searchField: searchField,
+              banner: banner,
               body: body,
             ),
           ),
@@ -145,24 +274,19 @@ Future<T?> showArnaPopupDialog<T>({
           transitionDuration: Styles.basicDuration,
           routeSettings: routeSettings,
           pageBuilder: (context, animation, secondaryAnimation) {
-            return ArnaPopupDialog(
-              title: title,
+            return _ArnaPopupDialog(
               headerBarLeading: headerBarLeading,
+              title: title,
               headerBarTrailing: headerBarTrailing,
               searchField: searchField,
+              banner: banner,
               body: body,
             );
           },
           transitionBuilder: (_, anim, __, child) => ScaleTransition(
-            scale: CurvedAnimation(
-              parent: anim,
-              curve: Styles.basicCurve,
-            ),
+            scale: CurvedAnimation(parent: anim, curve: Styles.basicCurve),
             child: FadeTransition(
-              opacity: CurvedAnimation(
-                parent: anim,
-                curve: Styles.basicCurve,
-              ),
+              opacity: CurvedAnimation(parent: anim, curve: Styles.basicCurve),
               child: child,
             ),
           ),